sevaflow/requirements.txt
# SEVAFLOW - GovTech Grievance Platform
# Python 3.10+

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Telegram Bot
python-telegram-bot>=20.7

# Database
aiosqlite>=0.19.0

# AI/LLM
google-generativeai>=0.3.2

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
